feat: TT-122 update tests
