+++ tests/utils/azure_users_test.py
@@ -1,5 +1,5 @@
 from unittest.mock import Mock, patch
-from utils.azure_users import AzureConnection, ROLE_FIELD_VALUES
+from utils.azure_users import AzureConnection, ROLE_FIELD_VALUES, AzureUser_v2
 from pytest import mark
 
 
@@ -40,3 +40,19 @@ def test_azure_connection_get_test_user_ids(get_mock):
     ids = ['ID1', 'ID2']
     az_conn = AzureConnection()
     assert az_conn.get_test_user_ids() == ids
+
+
+@patch('msal.ConfidentialClientApplication', Mock())
+@patch('utils.azure_users.AzureConnection.get_token', Mock())
+@patch('utils.azure_users.AzureConnection.get_test_user_ids')
+@patch('utils.azure_users.AzureConnection.users_v2')
+def test_azure_connection_get_non_test_users(
+    users_v2_mock, get_test_user_ids_mock
+):
+    test_user = AzureUser_v2('ID1', None, None, [])
+    non_test_user = AzureUser_v2('ID2', None, None, [])
+    users_v2_mock.return_value = [test_user, non_test_user]
+    get_test_user_ids_mock.return_value = ['ID1']
+    non_test_users = [non_test_user]
+    az_conn = AzureConnection()
+    assert az_conn.get_non_test_users() == non_test_users
